Adding doxygen comments

+++ Group-05/src/Parser.h
@@ -3,17 +3,20 @@
 
 #include <string>
 
+/**
+ * @brief Class to parse string equations
+ */
 class Parser {
 public:
   // Constructor
   Parser();
 
   // Member function to evaluate the expression
-  int Evaluate(std::string expression);
+  double Evaluate(std::string expression);
 
 private:
   // Helper function to parse a number inside curly braces
-  int ParseNumber(std::string expression, size_t &index);
+  double ParseNumber(std::string expression, size_t &index);
   auto MakeDivideFun(std::string name1, std::string name2);
   auto MakeAddFun(std::string name1, std::string name2);
   auto MakeSubtractFun(std::string name1, std::string name2);
